+++ R/outcome-estimates.R
@@ -73,10 +73,11 @@ nc_outcome_estimates <- function(.data, .graph, .outcome, .adjustment_vars, .mod
     all_top_models_tidied <- all_possible_models %>%
         # TODO: Have argument for threshold? For choosing number of models?
         map(~ MuMIn::get.models(.x, subset = TRUE)) %>%
-        # map(~ MuMIn::get.models(.x, subset = TRUE)) %>%
         imap_dfr(~ .tidy_all_model_outputs(.x, .y, .exponentiate = .exponentiate)) %>%
-        mutate(outcome = .outcome) %>%
-        select_at(vars("outcome", everything()))
+        mutate(outcome = .outcome,
+               adjusted_vars = paste(.adjustment_vars, collapse = ", ")) %>%
+        select_at(vars("outcome", everything())) %>%
+        dplyr::rename_all(~ gsub("\\.", "_", .))
 
     return(all_top_models_tidied)
 }